refactor(event): change `event.Bus` signature
`Subscribe` now returns two channels: an Event channel and an error
channel. The reason for this is that in most cases we want to handle
asynchronous errors, so they should be part of the Bus API. The
reasoning for not using the existing `Stream` type which provides both
the current Event and error comes from a purely personal preference. The
`Stream` types will likely be removed from the packages and replaced
with multiple return values (one output channel + one error channel).